requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pypdf==4.2.0
openpyxl==3.1.5
